+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-03-01
+
 ### ❗ Breaking Changes
 
 - Events are now defined by `GlyphButton.onMount/onShortPress/onLongPress/onTouchDown/onTouchUp` instead of `GlyphButton.addEventListener()`.
@@ -47,6 +49,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 Add section to make `release-it` not complain that this is missing. 2025-09-08 is when we really started working on this repository.
 
-[unreleased]: https://github.com/MissingCore/music-glyph-toys/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/MissingCore/music-glyph-toys/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/MissingCore/music-glyph-toys/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/MissingCore/music-glyph-toys/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/MissingCore/music-glyph-toys/releases/tag/v0.1.0